M4b step 1: trailer units and couplings, general static equilibrium, MAX_WHEELS 16
Trailer definitions (semitrailer, centre-axle, dolly drawbar) compose behind a towing vehicle
as units of one multibody tree on fifth-wheel/eye ball joints with roll springs and stops,
hinges and turntables; per-unit colliders and drag; energy-minimising static solver for any
layout (two-axle single units keep the lever-rule solver); test rigs settle, match statics,
roll straight and articulate in turns. Existing goldens unchanged.

+++ crates/control/src/ground/action.rs
@@ -86,7 +86,7 @@ pub enum PerWheelChannel {
     Throttle,
     Steering,
     /// Motor driving the wheels in the bit mask.
-    Drive(u8),
+    Drive(u32),
     Brake(usize),
     Steer(usize),
 }
